Moved guts of transfer functions to privates
Learned some things about mutability modifiers... this is all still pretty new.
Moved the guts of the transfer functions into their own private functions because calling externals internally was fixing to be a problem, and  repeating code is for noobs. 
Got rid of the variable mentioned in last change because I myself was being a noob, just passed "" to the other safeTransferFrom because duh.
Did some basic testing and everything seems to be working.